JOIN queries can be decomposed in to several separate SELECT statements.
SELECT a.name, b.job FROM Person a INNER JOIN Job b ON b.jobId = a.jobId WHERE a.age = 30
Can be separated in to two queries:

SELECT a.name FROM Person a WHERE a.age = 30
SELECT b.job FROM Job b WHERE b.jobId = ?
This is done in SelectStatement#prepare by checking the alias on QualifiedNames and ColumnMetadata and pulling selectors and relations in to separate SelectStatments.

As long as we only support AND operators we don't have to modify the original SelectStatment#prepare logic much.